requirements.txt
sentence-transformers>=2.2.2
jira>=3.5.1
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
